Show different buttons for logged out existing user (#99)
* remove navbar-inverse class to get rid of bootstrap effects

*  [USEC-993] Handle previously logged in users
`userPreviouslyLoggedIn` is a cookie which is set if user's previously logged-in or as registered. Currently we don't handle this case and thus we do not emphasize on login button.
Thus this PR emphasizes more on login instead of register buttons in header. When user is not previously logged-in; header items are as usual.

* refactor getItems to use object keys

* visual improvements

* add prop controls for previously logged in

* add autoprefixer

* remove comments

* refactor menu css

* refactor button classes